+++ CHANGELOG.md
@@ -10,8 +10,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   - `start_new_lines` the amount of new lines that should be printed before the conversation starts
   - `npc_text_fill_new_lines` should lined be filled between the NPC text and the player answer options
 - `world` condition now supports variables for the world name
-- `weather` condition now allows to define a world, which can be a variable and can now be used in static context
 - underscores are now escapable in item names and lore via `\_`
+- `weather` condition now allows to define a world, which can be a variable and can now be used in static context
 ### Changed
 - `spawn` event now only spawn mobs and no other entities
 ### Deprecated